added a bit more friction to delete a post. also fixed the verb issue in publishing/drafting posts

+++ public/scripts/dashboard.js
@@ -8,13 +8,15 @@ const deletePostEventHandler = (e) => {
   const postId = e.target.dataset.postId;
   const url = `/blog/dashboard/${postId}`;
 
+  if (!confirm("Are you sure you want to delete the blog?")) return;
+
   fetch(url, {
     method: "DELETE",
   })
     .then((res) => res.json())
     .then((data) => {
       if (data["status"] == "success") {
-        alert("successfully delete post");
+        alert("successfully deleted post");
         window.location.href = window.location.href;
       } else {
         alert("failed to delete post");
@@ -33,7 +35,11 @@ const publishDraftPostEventHandler = (e) => {
     .then((res) => res.json())
     .then((data) => {
       if (data["status"] == "success") {
-        alert("successfully toggled publish");
+        let msg = "published";
+        if (e.target.innerText === "[DRAFT]") {
+          msg = "drafted";
+        }
+        alert(`successfully ${msg} the post`);
         window.location.href = window.location.href;
       } else {
         alert("failed to toggle publish");
